requirements.txt
prefect==2.20.5
polars==0.20.31
duckdb==1.0.0
sqlalchemy==2.0.21
pydantic==2.5.3
fastapi==0.110.0
uvicorn==0.24.0
python-jose==3.3.0
python-multipart==0.0.7
plotly==5.17.0
python-dotenv==1.0.0
pyyaml==6.0.1
pytest
great-expectations